// EchoLog/src/generated/models/Cr4c3_rcasubmissionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Cr4c3_rcasubmissionsBase {
  createdon?: string;
  cr4c3_effectstatement?: string;
  importsequencenumber?: number;
  modifiedon?: string;
  cr4c3_rcasubmissionid?: string;
  cr4c3_rcatitle?: string;
  overriddencreatedon?: string;
  cr4c3_reviewcomments?: string;
  cr4c3_reviewedat?: string;
  statuscode?: number;
  cr4c3_status?: number;
  statecode?: number;
  cr4c3_submittedat?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
  _createdby_value?: string;
  _createdonbehalfby_value?: string;
  _createdonbehalfby_type?: _createdonbehalfby_type_createdonbehalfby_type;
  _createdby_type?: _createdby_type_createdby_type;
  _cr4c3_incident_value?: string;
  _cr4c3_incident_type?: _cr4c3_incident_type_cr4c3_incident_type;
  _modifiedby_value?: string;
  _modifiedonbehalfby_value?: string;
  _modifiedonbehalfby_type?: _modifiedonbehalfby_type_modifiedonbehalfby_type;
  _modifiedby_type?: _modifiedby_type_modifiedby_type;
  _ownerid_value?: string;
  _ownerid_type?: _ownerid_type_ownerid_type;
  _owningbusinessunit_value?: string;
  _owningbusinessunit_type?: _owningbusinessunit_type_owningbusinessunit_type;
  _cr4c3_reviewer_value?: string;
  _cr4c3_reviewer_type?: _cr4c3_reviewer_type_cr4c3_reviewer_type;
  _cr4c3_status_label?: _cr4c3_status_label_cr4c3_status_label;
  _statecode_label?: _statecode_label_statecode_label;
  _statuscode_label?: _statuscode_label_statuscode_label;
  _cr4c3_submittedby_value?: string;
  _cr4c3_submittedby_type?: _cr4c3_submittedby_type_cr4c3_submittedby_type;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
}