// EchoLog/src/generated/models/Cr4c3_incidentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Cr4c3_incidentsBase {
  cr4c3_createdat?: string;
  createdon?: string;
  cr4c3_description?: string;
  cr4c3_duedate?: string;
  importsequencenumber?: number;
  cr4c3_incidentid?: string;
  modifiedon?: string;
  overriddencreatedon?: string;
  cr4c3_rejectioncount?: number;
  cr4c3_severity?: number;
  statuscode?: number;
  cr4c3_status?: number;
  statecode?: number;
  cr4c3_ticketreference?: string;
  timezoneruleversionnumber?: number;
  cr4c3_title?: string;
  cr4c3_updatedat?: string;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
  _cr4c3_assignee_value?: string;
  _cr4c3_assignee_type?: _cr4c3_assignee_type_cr4c3_assignee_type;
  _createdby_value?: string;
  _createdonbehalfby_value?: string;
  _createdonbehalfby_type?: _createdonbehalfby_type_createdonbehalfby_type;
  _createdby_type?: _createdby_type_createdby_type;
  _cr4c3_department_value?: string;
  _cr4c3_department_type?: _cr4c3_department_type_cr4c3_department_type;
  _cr4c3_loggedby_value?: string;
  _cr4c3_loggedby_type?: _cr4c3_loggedby_type_cr4c3_loggedby_type;
  _modifiedby_value?: string;
  _modifiedonbehalfby_value?: string;
  _modifiedonbehalfby_type?: _modifiedonbehalfby_type_modifiedonbehalfby_type;
  _modifiedby_type?: _modifiedby_type_modifiedby_type;
  _ownerid_value?: string;
  _ownerid_type?: _ownerid_type_ownerid_type;
  _owningbusinessunit_value?: string;
  _owningbusinessunit_type?: _owningbusinessunit_type_owningbusinessunit_type;
  _cr4c3_process_value?: string;
  _cr4c3_process_type?: _cr4c3_process_type_cr4c3_process_type;
  _cr4c3_severity_label?: _cr4c3_severity_label_cr4c3_severity_label;
  _cr4c3_status_label?: _cr4c3_status_label_cr4c3_status_label;
  _statecode_label?: _statecode_label_statecode_label;
  _statuscode_label?: _statuscode_label_statuscode_label;
  _cr4c3_subdepartment_value?: string;
  _cr4c3_subdepartment_type?: _cr4c3_subdepartment_type_cr4c3_subdepartment_type;
  _cr4c3_team_value?: string;
  _cr4c3_team_type?: _cr4c3_team_type_cr4c3_team_type;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
}