// EchoLog/src/generated/models/Cr4c3_slarulesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Cr4c3_slarulesBase {
  createdon?: string;
  importsequencenumber?: number;
  cr4c3_isactive?: boolean;
  cr4c3_l1reviewpercent?: number;
  modifiedon?: string;
  overriddencreatedon?: string;
  cr4c3_severity?: number;
  cr4c3_slaname?: string;
  cr4c3_slaruleid?: string;
  statuscode?: number;
  statecode?: number;
  cr4c3_tathours?: number;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
  _createdby_value?: string;
  _createdonbehalfby_value?: string;
  _createdonbehalfby_type?: _createdonbehalfby_type_createdonbehalfby_type;
  _createdby_type?: _createdby_type_createdby_type;
  _modifiedby_value?: string;
  _modifiedonbehalfby_value?: string;
  _modifiedonbehalfby_type?: _modifiedonbehalfby_type_modifiedonbehalfby_type;
  _modifiedby_type?: _modifiedby_type_modifiedby_type;
  _ownerid_value?: string;
  _ownerid_type?: _ownerid_type_ownerid_type;
  _owningbusinessunit_value?: string;
  _owningbusinessunit_type?: _owningbusinessunit_type_owningbusinessunit_type;
  _cr4c3_severity_label?: _cr4c3_severity_label_cr4c3_severity_label;
  _statecode_label?: _statecode_label_statecode_label;
  _statuscode_label?: _statuscode_label_statuscode_label;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
}